+++ docs/features/exception-handling.md
@@ -6,10 +6,18 @@ This feature allows adding custom _Exception Handlers_.
 app.Features.AddExceptionHandling(...);
 ```
 
-## Default
+## Problem Details
 
-Adds default exception handler as well as middleware to log exceptions.
+Adds an exception handler implementation that returns errors in problem details
+format. It also adds a middleware that logs exceptions.
 
 ```csharp
-c => c.Default(typeUrlFormat: "https://my-service.com/errors/{0}")
+c => c.ProblemDetails(typeUrlFormat: "https://my-service.com/errors/{0}")
 ```
+
+> [!NOTE]
+>
+> More to find at [RFC 7807][rfc] and [Handle errors in ASP.NET Core][dotnet].
+
+[rfc]: https://www.rfc-editor.org/rfc/rfc7807.html
+[dotnet]: https://learn.microsoft.com/en-us/aspnet/core/fundamentals/error-handling?view=aspnetcore-8.0&source=recommendations#problem-details
