--- src/recipe/Baked.Recipe.Service.Application/ExceptionHandling/Default/DefaultExceptionHandlingExtensions.cs
@@ -1,12 +1,0 @@
-﻿using Baked.ExceptionHandling;
-using Baked.ExceptionHandling.Default;
-using Baked.Runtime;
-
-namespace Baked;
-
-public static class DefaultExceptionHandlingExtensions
-{
-    public static DefaultExceptionHandlingFeature Default(this ExceptionHandlingConfigurator _,
-        Setting<string>? typeUrlFormat = default
-    ) => new(typeUrlFormat);
-}
