+++ src/recipe/Baked.Recipe.Service.Application/ExceptionHandling/ProblemDetails/AuthenticationExceptionHandler.cs
@@ -1,7 +1,7 @@
 ﻿using System.Net;
 using System.Security.Authentication;
 
-namespace Baked.ExceptionHandling.Default;
+namespace Baked.ExceptionHandling.ProblemDetails;
 
 public class AuthenticationExceptionHandler : IExceptionHandler
 {
