+++ test/recipe/Baked.Test.Recipe.Service.Test/ExceptionHandling/HandlingExceptions.cs
@@ -1,5 +1,5 @@
 ﻿using Baked.ExceptionHandling;
-using Baked.ExceptionHandling.Default;
+using Baked.ExceptionHandling.ProblemDetails;
 using Baked.Orm;
 using Baked.Test.Orm;
 using System.Net;
